requirements.txt
streamlit==1.31.0
python-docx==1.0.1
pandas==2.2.0
python-dotenv==1.0.1